pyproject.toml
[project]
name = "mbg-stance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.185.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
]
